+++ src/app/static/src/app/hintVersion.ts
@@ -1,2 +1,3 @@
 
-export const currentHintVersion = "1.1.3";
+export const currentHintVersion = "1.6.0";
+