+++ src/app/static/src/tests/components/stepper.test.ts
@@ -310,7 +310,6 @@ describe("Stepper component", () => {
         expect(steps.at(1).props().active).toBe(true);
     });
 
-
     it("cannot go back from the first step", () => {
         const wrapper = createSut({country: "", ready: true}, {ready: true}, {}, {ready: true});
         const backLink = wrapper.find("#back");
@@ -603,5 +602,25 @@ describe("Stepper component", () => {
             {ready: true, result: "TEST" as any},
             {activeStep: step});
     };
+
+    it("complete step only becomes active/complete once state becomes ready", async () => {
+
+        const wrapper = createSut(completedBaselineState,
+            {},
+            {plottingMetadata: mockPlottingMetadataResponse()},
+            {ready: true},
+            {activeStep: 7});
+
+        let steps = wrapper.findAll(Step);
+        expect(steps.filter(s => s.props().active).length).toBe(0);
+
+        await makeReady(wrapper);
+        steps = wrapper.findAll(Step);
+        expect(steps.at(6).props().active).toBe(true);
+        expect(steps.at(6).props().complete).toBe(true);
+
+        const continueLink = wrapper.find("#continue")
+        expect(continueLink.classes()).toContain("disabled");
+    });
 
 });