+++ NEWS.md
@@ -1,3 +1,31 @@
+# hint 1.6.0
+
+* Show spinner when fetching ADR datasets
+
+# hint 1.5.1
+
+* Swap button colours on Save new version confirmation dialog
+
+# hint 1.5.0
+
+* Make 'Download results' step red when advance to that step.
+
+# hint 1.4.0
+
+* Include metadata default scale for bubble size
+
+# hint 1.3.0
+
+* Don't show ADR key to user after it has been entered.
+
+# hint 1.2.1
+
+* Fetch all relevant datasets from ADR without paging
+
+# hint 1.2.0
+
+* Support custom value ranges for bubble plot scales.
+
 # hint 1.1.3
 
 * Display text on map when no data in selections.